Broke down Utils class into separate classes according to their relations
