+++ app/src/main/java/com/ducttapeprogrammer/myapplication/utils/WeatherUtils.kt
@@ -1,15 +1,10 @@
-package com.ducttapeprogrammer.myapplication
+package com.ducttapeprogrammer.myapplication.utils
 
-import android.content.Context
 import android.graphics.drawable.Drawable
 import android.os.Build
 import android.widget.ImageView
 import androidx.annotation.RequiresApi
-import timber.log.Timber
-import java.text.ParseException
-import java.text.SimpleDateFormat
-import java.util.*
-import kotlin.math.roundToInt
+import com.ducttapeprogrammer.myapplication.*
 
 /**
  * This function will return the weather type ID providing the weatherCondition ID
@@ -168,115 +163,3 @@ fun ImageView.setRecyclerWeatherConditionIcon(weatherID: Int) {
 
 
 }
-
-
-/**
- * Set a SharedPreference for a Integer value
- */
-fun Int.setIntSharedPreference(sharedPrefKey: String) {
-
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    val editor = preference.edit()
-    editor.putInt(sharedPrefKey, this)
-    editor.apply()
-
-}
-
-/**
- * Get a SharedPreference for a Integer value
- */
-fun getIntSharedPreference(sharedPrefKey: String): Int {
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    return preference.getInt(sharedPrefKey, 0)
-}
-
-/**
- * Set a SharedPreference for a String value
- */
-fun String.setStringSharedPreference(sharedPrefKey: String) {
-
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    val editor = preference.edit()
-    editor.putString(sharedPrefKey, this)
-    editor.apply()
-
-}
-
-/**
- * Get a SharedPreference for a Integer value
- */
-fun getStringSharedPreference(sharedPrefKey: String): String? {
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    return preference.getString(sharedPrefKey, "")
-}
-
-/**
- * Set a SharedPreference for a boolean value
- */
-fun Boolean.setBooleanSharedPreference(sharedPrefKey: String) {
-
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    val editor = preference.edit()
-    editor.putBoolean(sharedPrefKey, this)
-    editor.apply()
-
-}
-
-/**
- * Get a SharedPreference for a Boolean value
- */
-fun getBooleanSharedPreference(sharedPrefKey: String): Boolean {
-    val preference = MyApplication.instance.applicationContext.getSharedPreferences(
-        MyApplication.instance.getString(R.string.app_name), Context.MODE_PRIVATE
-    )
-    return preference.getBoolean(sharedPrefKey, false)
-}
-
-
-/**
- * This function will convert the kelvin to degree celsius given the input
- * */
-fun convertKelvinToDegreeCelsius(kelvin: Double?): Int? {
-    return (kelvin?.toInt()?.minus(KELVIN))?.roundToInt()
-}
-
-
-/**
- * This function will return the date as a String. It will increment date by 1 day
- * every time this function is called. It uses a [SHARED_PREF_DATE_COUNT] to verify the
- * day is only incremented till 4 days, to make sure, only the next 4 days date is shown.
- * */
-fun getCurrentDate(): String {
-    if (getIntSharedPreference(SHARED_PREF_DATE_COUNT) > 4) {
-        val count = 0
-        count.setIntSharedPreference(SHARED_PREF_DATE_COUNT)
-    }
-    var count = getIntSharedPreference(SHARED_PREF_DATE_COUNT)
-    count += 1
-    count.setIntSharedPreference(SHARED_PREF_DATE_COUNT)
-    val oldDate: String? = SimpleDateFormat(DATE_FORMAT_ONE, Locale.getDefault()).format(Date())
-    val simpleDateFormat = SimpleDateFormat(DATE_FORMAT_ONE, Locale.getDefault())
-    val calendar = Calendar.getInstance()
-    try {
-        calendar.time = simpleDateFormat.parse(oldDate)
-    } catch (e: ParseException) {
-        e.printStackTrace()
-    }
-    calendar.add(Calendar.DAY_OF_MONTH, count)
-
-    return simpleDateFormat.format(calendar.time)
-}
-
-
-
-
